bot/requirements.txt
aiogram==3.15.0
aiohttp==3.11.12
python-dotenv==1.0.1
